# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CasperHolders API",
    "version": "1.0.0"
  },
  "paths": {
    "/apy/current": {
      "get": {
        "description": "Return current APY",
        "responses": {
          "200": {
            "description": "Current APY",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/APY"
                }
              }
            }
          },
          "503": {
            "description": "No server available"
          }
        }
      }
    },
    "/deploys/{hash}": {
      "get": {
        "description": "Return a deploy if present in the database",
        "parameters": [
          {
            "in": "path",
            "name": "hash",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Hash of the deploy"
          }
        ],
        "responses": {
          "200": {
            "description": "Return a deploy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Deploy"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/deploys/": {
      "post": {
        "description": "Return a deploy if present in the database",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "deploy",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Deploy"
            },
            "required": true,
            "description": "Deploy json object"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Deploy already exist or deploy malformed"
          },
          "503": {
            "description": "Couldn't insert the deploy in the database"
          }
        }
      }
    },
    "/validators/accountinfos": {
      "get": {
        "description": "Get validator info from the blockchain auction state. Add additional metadata asynchronously",
        "responses": {
          "200": {
            "description": "Return array of validator infos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/ValidatorsInfos"
                }
              }
            }
          },
          "503": {
            "description": "No server available"
          }
        }
      }
    }
  },
  "definitions": {
    "APY": {
      "type": "number"
    },
    "Deploy": {
      "type": "object",
      "properties": {
        "deploy": {
          "type": "object",
          "required": true
        }
      }
    },
    "ValidatorInfo": {
      "type": "object",
      "required": [
        "name",
        "group",
        "delegation_rate",
        "publicKey",
        "staked_amount"
      ],
      "properties": {
        "name": {
          "type": "string",
          "description": "Name or public key of the validator"
        },
        "group": {
          "type": "string",
          "description": "Group of the validator Active or Inactive"
        },
        "delegation_rate": {
          "type": "int",
          "description": "Delegation rate of the validator"
        },
        "publicKey": {
          "type": "boolean",
          "description": "Public key of the validator"
        },
        "staked_amount": {
          "type": "string",
          "description": "Number of CSPR staked on the validator"
        }
      },
      "example": {
        "name": "CasperHolders",
        "group": "Active",
        "delegation_rate": 0,
        "publicKey": "0124bfdae2ed128fa5e4057bc398e4933329570e47240e57fc92f5611a6178eba5",
        "staked_amount": "100"
      }
    },
    "ValidatorsInfos": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/ValidatorInfo"
      }
    }
  },
  "components": {},
  "tags": []
}